fix(boft, oft): preserve half-precision dtype in MultiplicativeDropoutLayer
`MultiplicativeDropoutLayer.forward` in both BOFT and OFT builds its mask and
identity-matrix buffers with `device=x.device` but no `dtype=`, so they default
to float32. Combined with a float16/bfloat16 input via ordinary arithmetic
(`(1 - mask) * x + mask * eye_matrix`), the result silently promotes to
float32 whenever dropout is actually applied during training (self.training
and, for OFT, p > 0).

Verified directly against the real classes:
  BOFT: float16 -> float32, bfloat16 -> float32 (unfixed)
  OFT:  float16 -> float32, bfloat16 -> float32 (unfixed)
Fixed by adding dtype=x.dtype alongside device=x.device on every buffer built
in these two forward() methods. Confirmed the fixed float16 output is
numerically identical (torch.equal) to a float32 reference run through the
same code with the same random state -- only the dtype changes, not the
values (mask/eye entries are exact 0/1 in both dtypes).

Adds a parametrized regression test to each: fails on the current code (4/4
failures reverting only the production files), passes with the fix (4/4).
Full test_boft.py + test_oft.py: 7 passed. ruff check/format clean on the
touched lines (repo pins ruff==0.16.4; checked against the real project
config, not --isolated, by temporarily dropping the version pin locally and
reverting it before this commit -- not part of this diff).

+++ src/peft/tuners/boft/layer.py
@@ -180,16 +180,21 @@ class MultiplicativeDropoutLayer(nn.Module):
             num_zeros = D - num_to_replace
 
             # Generate a flat tensor with desired number of 1s and 0s
-            mask = torch.cat([torch.ones(num_to_replace, device=x.device), torch.zeros(num_zeros, device=x.device)])
+            mask = torch.cat(
+                [
+                    torch.ones(num_to_replace, device=x.device, dtype=x.dtype),
+                    torch.zeros(num_zeros, device=x.device, dtype=x.dtype),
+                ]
+            )
 
             # Shuffle and reshape the mask
             mask = mask[torch.randperm(D)].view(1, D, 1, 1)
 
-            full_mask = torch.zeros(N, D, 1, 1, device=x.device)
+            full_mask = torch.zeros(N, D, 1, 1, device=x.device, dtype=x.dtype)
             full_mask[n_random] = mask
 
             # Use the mask to combine original matrices and identity matrices
-            eye_matrix = torch.eye(H, device=x.device).repeat(N, D, 1, 1)
+            eye_matrix = torch.eye(H, device=x.device, dtype=x.dtype).repeat(N, D, 1, 1)
             x = (1 - full_mask) * x + full_mask * eye_matrix
         return x
